working notifier system
