+++ pkg/sync/kubernetes/featureflagconfiguration/clientset.go
@@ -1,0 +1,75 @@
+package featureflagconfiguration
+
+import (
+	"fmt"
+	"time"
+
+	ffv1alpha1 "github.com/open-feature/open-feature-operator/apis/core/v1alpha1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/apimachinery/pkg/watch"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/cache"
+)
+
+type FFCInterface interface {
+	FeatureFlagConfigurations(namespace string) FeatureFlagConfigurationInterface
+}
+
+type FFCClient struct {
+	restClient rest.Interface
+}
+
+func WatchResources(clientSet FFCInterface) cache.Store {
+	projectStore, projectController := cache.NewInformer(
+		&cache.ListWatch{
+			ListFunc: func(lo metav1.ListOptions) (result runtime.Object, err error) {
+				return clientSet.FeatureFlagConfigurations("*").List(lo)
+			},
+			WatchFunc: func(lo metav1.ListOptions) (watch.Interface, error) {
+				return clientSet.FeatureFlagConfigurations("*").Watch(lo)
+			},
+		},
+		&ffv1alpha1.FeatureFlagConfiguration{},
+		1*time.Minute,
+		cache.ResourceEventHandlerFuncs{
+			AddFunc: func(obj interface{}) {
+				fmt.Printf("service added: %s \n", obj)
+			},
+			DeleteFunc: func(obj interface{}) {
+				fmt.Printf("service deleted: %s \n", obj)
+			},
+			UpdateFunc: func(oldObj, newObj interface{}) {
+				fmt.Printf("service changed \n")
+			},
+		},
+	)
+
+	go projectController.Run(wait.NeverStop)
+	return projectStore
+}
+
+func NewForConfig(c *rest.Config) (*FFCClient, error) {
+	config := *c
+	config.ContentConfig.GroupVersion = &schema.GroupVersion{Group: ffv1alpha1.GroupVersion.Group, Version: ffv1alpha1.GroupVersion.Version}
+	config.APIPath = "/apis"
+	config.UserAgent = rest.DefaultKubernetesUserAgent()
+	config.NegotiatedSerializer = serializer.NewCodecFactory(scheme.Scheme)
+	client, err := rest.RESTClientFor(&config)
+	if err != nil {
+		return nil, err
+	}
+
+	return &FFCClient{restClient: client}, nil
+}
+
+func (c *FFCClient) FeatureFlagConfigurations(namespace string) FeatureFlagConfigurationInterface {
+	return &FeatureFlagClient{
+		restClient: c.restClient,
+		ns:         namespace,
+	}
+}
